Add "Priority Speaker" permission as constant (#628)
* Bump to v0.17.0

* Add members from GuildMembersChunk to state (#454)

* Revert "Add members from GuildMembersChunk to state (#454)" (#455)

This reverts commit f38d2654c1c8c1814487cf55296331d98f23faac.

* added clarification when initializing discordgo

from the code I ran on my system with the latest changes this seems to be the syntax for the authentication tokens. I'm guessing it was just never updated.
Let me know if this is incorrect. 
Thanks!

* travis: update go versions

* Add "Priority Speaker" permission constant

The permission "Priority Speaker" was missing as constant. Permission has been added as constant and to the "PermissionAllVoice" constant as well.